[progress] refactored megamenu component structure
